+++ Coursera_Task/Data/Models/Report.cs
@@ -7,8 +7,8 @@
         public string LastName { get; set; }
         public string CourseName { get; set; }
         public DateTime CompletionDate { get; set; }
-        public int TotalTime { get; set; }
-        public int Credit { get; set; }
+        public byte TotalTime { get; set; }
+        public byte Credit { get; set; }
         public string InstructorFirstName { get; set; }
         public string InstructorLastName { get; set; }
     }
